+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - lookup button form element
+- Rendering of form `footer` HTML
 
 ## [8.6.0] - 2025-07-01